bq-analytics-agent/requirements.txt
# Google Cloud SDKs
google-cloud-geminidataanalytics>=0.1.0
google-cloud-storage>=2.14.0
google-cloud-dataplex>=1.12.0

# Protobuf / gRPC
proto-plus>=1.22.0
protobuf>=4.25.0
google-auth>=2.27.0
grpcio>=1.60.0

# Streamlit UI
streamlit>=1.31.0

# Data & Visualization
pandas>=2.1.0
altair>=5.2.0

# Config & Validation
pyyaml>=6.0
pydantic>=2.5.0

# Code display
pygments>=2.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
